Move bucket creation attributes one level up to allow bucket createion in Desktop client.
