+++ backend/src/main/java/org/cryptomator/hub/api/cipherduck/StorageProfileS3STSDto.java
@@ -10,48 +10,7 @@ import java.util.UUID;
 
 public final class StorageProfileS3STSDto extends StorageProfileS3Dto {
 
-	public enum S3_SERVERSIDE_ENCRYPTION {
-		NONE, SSE_AES256, SSE_KMS_DEFAULT
-	}
 
-	//======================================================================
-	// (2) STS only: bucket creation
-	//======================================================================
-	@JsonProperty(value = "region", required = true, defaultValue = "us-east-1")
-	@Schema(description = "Default region selected in the frontend/client to create bucket in.", example = "443", defaultValue = "us-east-1")
-	String region = "us-east-1";
-
-	@JsonProperty(value = "regions", required = true)
-	@Schema(description = "List of selectable regions in the frontend/client to create bucket in. Defaults to full list from AWS SDK.")
-	List<String> regions = Region.regions().stream().map(Region::id).toList();
-
-	@JsonProperty(value = "bucketPrefix", required = true)
-	@Schema(description = "Buckets are create with name <bucket prefix><vault UUID>.", example = "cipherduck")
-	String bucketPrefix;
-
-	@JsonProperty(value = "stsRoleArnClient", required = true)
-	@Schema(description = "STS role for clients to assume to create buckets. Will be the same as stsRoleArnHub for AWS, different for MinIO.", example = "arn:aws:iam::<ACCOUNT ID>:role/cipherduck-createbucket")
-	String stsRoleArnClient;
-
-	@JsonProperty(value = "stsRoleArnHub", required = true)
-	@Schema(description = "STS role for frontend to assume to create buckets (used with inline policy and passed to hub storage). Will be the same as stsRoleArnClient for AWS, different for MinIO.", example = "arn:aws:iam::<ACCOUNT ID>:role/cipherduck-createbucket")
-	String stsRoleArnHub;
-
-	@JsonProperty("stsEndpoint")
-	@Schema(description = "STS endpoint to use for AssumeRoleWithWebIdentity and AssumeRole for getting a temporary access token passed to the storage. Defaults to AWS SDK default.", nullable = true)
-	String stsEndpoint;
-
-	@JsonProperty(value = "bucketVersioning", defaultValue = "true", required = true)
-	@Schema(description = "Enable bucket versioning upon bucket creation", defaultValue = "true", required = true)
-	Boolean bucketVersioning = true;
-
-	@JsonProperty(value = "bucketAcceleration")
-	@Schema(description = "Enable bucket versioning upon bucket creation (null for MinIO)", nullable = true)
-	Boolean bucketAcceleration = null;
-
-	@JsonProperty(value = "bucketEncryption", required = true)
-	@Schema(description = "Enable bucket versioning upon bucket creation", required = true)
-	S3_SERVERSIDE_ENCRYPTION bucketEncryption = S3_SERVERSIDE_ENCRYPTION.NONE;
 
 	//----------------------------------------------------------------------
 	// (3b) STS client profile custom properties
@@ -138,42 +97,6 @@ public final class StorageProfileS3STSDto extends StorageProfileS3Dto {
 		storageProfile.stsRoleArn2 = this.stsRoleArn2;
 		storageProfile.stsDurationSeconds = this.stsDurationSeconds;
 		return storageProfile;
-	}
-
-	public String region() {
-		return region;
-	}
-
-	public List<String> regions() {
-		return regions;
-	}
-
-	public String bucketPrefix() {
-		return bucketPrefix;
-	}
-
-	public String stsRoleArnClient() {
-		return stsRoleArnClient;
-	}
-
-	public String stsRoleArnHub() {
-		return stsRoleArnHub;
-	}
-
-	public String stsEndpoint() {
-		return stsEndpoint;
-	}
-
-	public Boolean bucketVersioning() {
-		return bucketVersioning;
-	}
-
-	public Boolean bucketAcceleration() {
-		return bucketAcceleration;
-	}
-
-	public S3_SERVERSIDE_ENCRYPTION bucketEncryption() {
-		return bucketEncryption;
 	}
 
 	public String stsRoleArn() {
