+++ backend/src/main/java/org/cryptomator/hub/entities/cipherduck/StorageProfileS3STS.java
@@ -12,36 +12,6 @@ import java.util.List;
 @DiscriminatorValue("S3STS")
 public class StorageProfileS3STS extends StorageProfileS3 { // TODO make sealed/final?
 
-	//======================================================================
-	// (2) STS only: bucket creation
-	//======================================================================
-	@Column
-	public String region;
-
-	@Column
-	public List<String> regions;
-
-	@Column
-	public String bucketPrefix;
-
-	@Column
-	public String stsRoleArnClient;
-
-	@Column
-	public String stsRoleArnHub;
-
-	@Column
-	public String stsEndpoint = null;
-
-	@Column
-	public Boolean bucketVersioning = true;
-
-	@Column
-	public Boolean bucketAcceleration = true;
-
-	@Column
-	public String bucketEncryption;
-
 	//----------------------------------------------------------------------
 	// (3b) STS client profile custom properties
 	//----------------------------------------------------------------------
